+++ CHANGES.md
@@ -68,6 +68,8 @@ To be released.
      -  Added `VerifyRequestOptions.tracerProvider` option.
      -  Added `SignRequestOptions` interface.
      -  Added the optional fourth parameter to `signRequest()` function.
+     -  Added `VerifyProofOptions.tracerProvider` option.
+     -  Added `VerifyObjectOptions.tracerProvider` option.
 
  -  Added `@fedify/fedify/x/sveltekit` module for integrating with [SvelteKit]
     hook.  [[#171], [#183] by Jiyu Park]